requirements.txt
dj-database-url~=0.5.0
Django~=2.0.0
django-configurations~=2.0
Pillow~=5.2.0
psycopg2~=2.7.0
python-dotenv~=0.8.0
pytz==2018.5
PyYAML~=3.0
django-registration-redux~=2.0
